Extract `Applications` and CSV-parsing functions to own module
I had to create  `util` module as well
