+++ grader/test_applications.py
@@ -4,7 +4,8 @@ from textwrap import dedent
 from pytest import raises
 
 from configfile import ConfigFile
-from grader import Applications, build_person_factory, list_of_str
+from applications import Applications, build_person_factory
+from util import list_of_str
 
 
 def _tmp_application_files(tmpdir, config_string, csv_string):
